+++ instrumentation/opentelemetry-instrumentation-django/src/opentelemetry/instrumentation/django/middleware.py
@@ -12,7 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import time
 from logging import getLogger
+
+from django.conf import settings
 
 from opentelemetry.configuration import Configuration
 from opentelemetry.context import attach, detach
@@ -41,11 +44,16 @@ except ImportError:
     MiddlewareMixin = object
 
 _logger = getLogger(__name__)
+_attributes_by_preference = [
+    ["http.scheme", "http.host", "http.target"],
+    ["http.scheme", "http.server_name", "net.host.port", "http.target"],
+    ["http.scheme", "net.host.name", "net.host.port", "http.target"],
+    ["http.url"],
+]
 
 
 class _DjangoMiddleware(MiddlewareMixin):
-    """Django Middleware for OpenTelemetry
-    """
+    """Django Middleware for OpenTelemetry"""
 
     _environ_activation_key = (
         "opentelemetry-instrumentor-django.activation_key"
@@ -88,6 +96,21 @@ class _DjangoMiddleware(MiddlewareMixin):
         except Resolver404:
             return "HTTP {}".format(request.method)
 
+    @staticmethod
+    def _get_metric_labels_from_attributes(attributes):
+        labels = {}
+        labels["http.method"] = attributes.get("http.method", "")
+        for attrs in _attributes_by_preference:
+            labels_from_attributes = {
+                attr: attributes.get(attr, None) for attr in attrs
+            }
+            if set(attrs).issubset(attributes.keys()):
+                labels.update(labels_from_attributes)
+                break
+        if attributes.get("http.flavor"):
+            labels["http.flavor"] = attributes.get("http.flavor")
+        return labels
+
     def process_request(self, request):
         # request.META is a dictionary containing all available HTTP headers
         # Read more about request.META here:
@@ -95,6 +118,9 @@ class _DjangoMiddleware(MiddlewareMixin):
 
         if self._excluded_urls.url_disabled(request.build_absolute_uri("?")):
             return
+
+        # pylint:disable=W0212
+        request._otel_start_time = time.time()
 
         environ = request.META
 
@@ -110,8 +136,13 @@ class _DjangoMiddleware(MiddlewareMixin):
             ),
         )
 
+        attributes = collect_request_attributes(environ)
+        # pylint:disable=W0212
+        request._otel_labels = self._get_metric_labels_from_attributes(
+            attributes
+        )
+
         if span.is_recording():
-            attributes = collect_request_attributes(environ)
             attributes = extract_attributes_from_object(
                 request, self._traced_request_attrs, attributes
             )
@@ -176,6 +207,10 @@ class _DjangoMiddleware(MiddlewareMixin):
                 "{} {}".format(response.status_code, response.reason_phrase),
                 response,
             )
+            # pylint:disable=W0212
+            request._otel_labels["http.status_code"] = str(
+                response.status_code
+            )
             request.META.pop(self._environ_span_key)
 
             request.META[self._environ_activation_key].__exit__(
@@ -186,5 +221,15 @@ class _DjangoMiddleware(MiddlewareMixin):
         if self._environ_token in request.META.keys():
             detach(request.environ.get(self._environ_token))
             request.META.pop(self._environ_token)
+
+        try:
+            metric_recorder = getattr(settings, "OTEL_METRIC_RECORDER", None)
+            if metric_recorder is not None:
+                # pylint:disable=W0212
+                metric_recorder.record_server_duration_range(
+                    request._otel_start_time, time.time(), request._otel_labels
+                )
+        except Exception as ex:  # pylint: disable=W0703
+            _logger.warning("Error recording duration metrics: %s", ex)
 
         return response