+++ instrumentation/opentelemetry-instrumentation-django/CHANGELOG.md
@@ -11,6 +11,8 @@ Released 2020-10-13
 - Changed span name extraction from request to comply semantic convention ([#992](https://github.com/open-telemetry/opentelemetry-python/pull/992)) 
 - Added support for `OTEL_PYTHON_DJANGO_TRACED_REQUEST_ATTRS` ([#1154](https://github.com/open-telemetry/opentelemetry-python/pull/1154))
 - Added capture of http.route ([#1226](https://github.com/open-telemetry/opentelemetry-python/issues/1226))
+- Add support for tracking http metrics
+  ([#1230](https://github.com/open-telemetry/opentelemetry-python/pull/1230))
 
 ## Version 0.13b0